cubes/gaetandi/project/requirements.txt
easydict==1.9
PyYAML
typer
torch
torchvision
pandas
tqdm
